Added save guard and fixed local save data bug

+++ js/List.js
@@ -9,7 +9,7 @@ function addRow() {
   var row = table.insertRow(rowCount);
   row.insertCell(0).innerHTML= listItem.value;
   row.insertCell(1).innerHTML= 
-  '<svg xmlns="http://www.w3.org/2000/svg" onClick="deleteRow(this)" class="icon icon-tabler icon-tabler-trash" width="52" height="52" viewBox="0 0 24 24" stroke-width="1.5" stroke="#F44336" fill="none" stroke-linecap="round" stroke-linejoin="round" style="float: right;"><path stroke="none" d="M0 0h24v24H0z"/><line x1="4" y1="7" x2="20" y2="7" /><line x1="10" y1="11" x2="10" y2="17" /><line x1="14" y1="11" x2="14" y2="17" /><path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" /><path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" /></svg>';
+  '<svg xmlns="http://www.w3.org/2000/svg" onClick="deleteRow(this)" class="icon icon-tabler icon-tabler-trash list-delete-cell" width="52" height="52" viewBox="0 0 24 24" stroke-width="1.5" stroke="#F44336" fill="none" stroke-linecap="round" stroke-linejoin="round" style="float: right;"><path stroke="none" d="M0 0h24v24H0z"/><line x1="4" y1="7" x2="20" y2="7" /><line x1="10" y1="11" x2="10" y2="17" /><line x1="14" y1="11" x2="14" y2="17" /><path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" /><path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" /></svg>';
   document.getElementById("listItem").value="";
 }
  
@@ -20,17 +20,23 @@ function deleteRow(obj) {
 }
 
 function save() {
-  var listToSave = document.getElementById("display-table").innerHTML;
-  var d = new Date();
-  lastSaveDate = d.toLocaleString();
-  localStorage.setItem('SavedList', listToSave);
-  localStorage.setItem('LastSaveDate', lastSaveDate);
-  // sets the date on screen
-  document.getElementById('lastSaveDate').innerHTML = lastSaveDate;
-  // snackbar DIV
-  var x = document.getElementById("saveSnackbar");
-  x.className = "show";
-  setTimeout(function(){ x.className = x.className.replace("show", ""); }, 2000);
+  var r = confirm("Overwrite save data?");
+  if (r == true) {
+    var listToSave = document.getElementById("display-table").innerHTML;
+    var d = new Date();
+    listLastSaveDate = d.toLocaleString();
+    localStorage.setItem('SavedList', listToSave);
+    localStorage.setItem('ListLastSaveDate', listLastSaveDate);
+    // sets the date on screen
+    document.getElementById('listLastSaveDate').innerHTML = listLastSaveDate;
+    // snackbar DIV
+    var x = document.getElementById("saveSnackbar");
+    x.className = "show";
+    setTimeout(function(){ x.className = x.className.replace("show", ""); }, 2000);
+  } else {
+    return;
+  }
+  
 }
 
 function load() {
@@ -48,8 +54,8 @@ function load() {
 }
 
 function checkSaveList() {
-  var savedLastSaveDate = localStorage.getItem('LastSaveDate');
+  var savedLastSaveDate = localStorage.getItem('ListLastSaveDate');
   if (savedLastSaveDate) {
-    document.getElementById('lastSaveDate').innerHTML = savedLastSaveDate;
+    document.getElementById('listLastSaveDate').innerHTML = savedLastSaveDate;
   }
 }
